Remove RunFromPackage setting for Github deployment
