+++ appservice/package.json
@@ -1,7 +1,7 @@
 {
     "name": "vscode-azureappservice",
     "author": "Microsoft Corporation",
-    "version": "0.29.2",
+    "version": "0.29.3",
     "description": "Common tools for developing Azure App Service extensions for VS Code",
     "tags": [
         "azure",
